Prepare for atecc608a driver implementation
Add psa-crypto as a git submodule in order to use the psa-crypto
headers files.

Add an mbedignore file in order to ignore the psa-crypto submodule
except for the psa-crypto header files.

Add an mbed_lib.json file in order to use the default mbedtls
config.h file used by mbed-os.